All changes until 05/23/2016
Master commit, version 1.0.0 build 1.0.7 (final) completed.

- Pod updates and modifications.
- Storyboard layout bugs and issues
- Firebase implemented
- Event adding bug regarding projects and displays
- Improved stability
- Other small changes

Notes:
Version has been submitted to the App Store.

+++ Done Widget/TodayViewController.m
@@ -44,7 +44,7 @@
     [formate setDateFormat:@"yyyy-MM-dd HH:mm"];
     self.upComingDate.text = [formate stringFromDate:date];
     self.upComingTitle.text = title;
-    self.todaySummary.text = [NSString stringWithFormat:@"There are %lu events today and you have completed %lu of them.", totalEvents.integerValue, completedEvents.integerValue];
+    self.todaySummary.text = [NSString stringWithFormat:@"There are %lu events today and you have completed %lu of them.", (long)totalEvents.integerValue, (long)completedEvents.integerValue];
 }
 
 - (void)didReceiveMemoryWarning {
